+++ Core/OfficeDevPnP.Core/Properties/AssemblyInfo.cs
@@ -16,7 +16,7 @@ using System.Runtime.InteropServices;
 #endif
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("OfficeDevPnP.Core")]
-[assembly: AssemblyCopyright("Copyright © Microsoft 2018")]
+[assembly: AssemblyCopyright("Copyright © Microsoft 2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: NeutralResourcesLanguage("en-US")]
@@ -43,7 +43,7 @@ using System.Runtime.InteropServices;
 // Minor version = Sequence...version 0 was with Aug 2018 release, 1 = Sept, 2 = Oct,...
 // Third part = version indenpendant showing the release month in YYMM
 // Fourth part = 0 normally or a sequence number when we do an emergency release
-[assembly: AssemblyVersion("3.5.1901.0")]
-[assembly: AssemblyFileVersion("3.5.1901.0")]
+[assembly: AssemblyVersion("3.6.1902.0")]
+[assembly: AssemblyFileVersion("3.6.1902.0")]
 
 [assembly: InternalsVisibleTo("OfficeDevPnP.Core.Tests")]